% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAST.R
\name{bh.func}
\alias{bh.func}
\title{BH procedure}
\usage{
bh.func(pv, q)
}
\arguments{
\item{pv}{p-values}

\item{q}{nominal FDR level}
}
\description{
This function runs the BH procedure, it is used for screening purposes only in the offline threshold determination step.
}
