% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAST.R
\name{DenEst}
\alias{DenEst}
\title{Estimate density from observations}
\usage{
DenEst(x.data, x.eval)
}
\arguments{
\item{x.data}{data points for density estimation}

\item{x.eval}{evaluation points for estimated density}
}
\description{
This function estimate density from a set of observations and evalute at provided data points.
}
