% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAST.R
\name{gamma_calc}
\alias{gamma_calc}
\title{Calculate offline procedure for determing screening threshold}
\usage{
gamma_calc(x.Lfdr, t, N = 200, alpha = 0.05)
}
\arguments{
\item{x.Lfdr}{estimated test statistic (conditional local-false discovery rate)}

\item{t}{current testing location}

\item{N}{neighborhood size}

\item{alpha}{nominal FDR level, default is 0.05}
}
\description{
This function runs the offline procedure for determining a dynamic threshold for screening test statistics CLfdr.
}
