% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAST.R
\name{lin.itp}
\alias{lin.itp}
\title{Linear Interpolation}
\usage{
lin.itp(x, X, Y)
}
\arguments{
\item{x}{the coordinates of points where the density needs to be interpolated}

\item{X}{the coordinates of the estimated densities}

\item{Y}{the values of the estimated densities}
}
\description{
This function returns the linearly interpolated densities.
}
