% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAST.R
\name{vary_p.Est}
\alias{vary_p.Est}
\title{Estimate varying proportion of non-null signals}
\usage{
vary_p.Est(z_score, init = 200, N, h)
}
\arguments{
\item{z_score}{standardized observations}

\item{init}{initial burn in period of ovservations. Default is 200.}

\item{N}{neighborhood size}

\item{h}{bandwidth choice}
}
\description{
This function estimates the non-null proportion of signals using conditional density estimator with screening on p-values.
}
